Modified normalized paths
