+++ mango/tests/test_file_functions.py
@@ -2,9 +2,9 @@ from unittest import TestCase
 
 from mango.processing.file_functions import (
     list_files_directory,
-    load_json,
-    normalize_path,
+    load_json
 )
+from mango.tests.const import normalize_path
 
 
 class FileTests(TestCase):
